Refactor destination list into separate component
See DIH-241
