+++ src/router.jsx
@@ -17,6 +17,7 @@ import DestinationVolunteers from './sections/admin/destinations/destination/vol
 import DestinationEmails from './sections/admin/destinations/destination/emailTemplates';
 import DestinationAddVolunteer from './sections/admin/destinations/destination/addVolunteer';
 import Destinations from './sections/admin/destinations';
+import DestinationsList from './sections/admin/destinations/destinationsList';
 import SignupTrip from './sections/trips/signup';
 import Trips from './sections/trips/';
 import Trip from './sections/trips/trip';
@@ -50,7 +51,9 @@ export default(
                     <Route name="Trip" path="/trips/:tripId" component={Trip} />
                 </Route>
                 <Route name="Destinations" path="admin/destinations">
-                    <IndexRoute component={Destinations} />
+                    <Route component={Destinations}>
+                        <IndexRoute component={DestinationsList} />
+                    </Route>
                     <Route path=":destinationId" component={Destination}>
                         <IndexRoute component={DestinationVolunteers} />
                         <Route name="Emails" path="emails" component={DestinationEmails} />
