+++ src/sections/admin/destinations/destinationsList.jsx
@@ -1,0 +1,44 @@
+import React, { PropTypes, Component } from 'react';
+import { connect } from 'react-redux';
+import { list } from '../../../actions/destinationActions';
+import Table from '../../../commons/Table';
+
+const createHandlers = (dispatch) => () => dispatch(list());
+
+class DestinationsList extends Component {
+    constructor(props) {
+        super(props);
+        this.handlers = createHandlers(this.props.dispatch);
+    }
+
+    componentDidMount() {
+        this.handlers();
+    }
+
+    render() {
+        return (
+            <Table
+                columnNames={{
+                    name: 'Name'
+                }}
+                itemKey="id"
+                link={{
+                    columnName: 'name',
+                    prefix: '/admin/destinations/'
+                }}
+                items={this.props.destinations}
+            />
+        );
+    }
+}
+
+const mapStateToProps = store => ({
+    destinations: store.destinationState.destinations
+});
+
+DestinationsList.propTypes = {
+    dispatch: PropTypes.func.isRequired,
+    destinations: PropTypes.array.isRequired
+};
+
+export default connect(mapStateToProps)(DestinationsList);
